install: Clean stale http download files
If the download is interrupted the temproary file will be
left in the /slot and it will fill up and then fail. This
will share the secret between network and install and could
be done better by the upstream devs.

I don't know how g_dir_read_name works with NULL so I added
a NULL check.