install: Go from overlay.tar.bz2 to not having one
In case a overlay.tar.bz2 is used for debugging and then removed,
it would remain in the slot and the system would behave different
than expected.

Now we don't want to delete all the files (e.g. the manifest or
the slot state) so let's treat the overlay.tar.bz2 specially as
we know how it is being used.

Upstream might have a blacklist or whitelist of files to keep even
if they were not in the manifest.