install: Do not accidently pass on failure
In case the slot is already/still mounted the update will fail
but claim a success. The issue is that if no mountpoint has been
returned it would return on the failure path with a positive result.

Try to make the code a bit more robust and check if we have at
least written the thing once.